stop using config_types
